more properly handle long periods of vicon drops
We will still add states during ~10sec drops in vicon readings.
But during these times, we will not use any vicon measurements as factors.
This means just the IMU is used to recover the states during these times.
We hard enforce that there is a vicon pose with in 0.2 sec to add as a valid factor.
This makes sure that our linear interpolation is an ok approximation (but we could be stricter...).